+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Enable passing `string`s to `addPeerToAddressBook`.
+- Use `addPeerToAddressBook` in examples.
+
 ## [0.1.0] - 2021-05-12
 
 ### Added